backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
sqlalchemy 
Flask-JWT-Extended
psycopg2-binary
requests
bcrypt
python-dotenv
email-validator
Werkzeug